BSSeq Mixin
